requirements.txt
langchain
langchain-community
google-genai
sentence-transformers
pytesseract
Pillow
pymupdf
indic-transliteration
langdetect
tiktoken
python-dotenv
fastapi
uvicorn
streamlit
chromadb
torchvision
